+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * Fix case-sensitive tests. Fixes UIIN-329.
 * Add electronic access relationship reference data (UIIN-316)
+* Remove notes helper app
 
 ## [1.3.0](https://github.com/folio-org/ui-inventory/tree/v1.3.0) (2018-09-27)
 [Full Changelog](https://github.com/folio-org/ui-inventory/compare/v1.2.1...v1.3.0)