+++ CHANGELOG.md
@@ -2,8 +2,9 @@
 
 ## 1.4.0 (IN PROGRESS)
 
+* Remove `instance.instanceFormatId`, add `instance.instanceFormatIds` (UIIN-330)
 * Fix case-sensitive tests. Fixes UIIN-329.
-* Remove notes helper app
+* Remove notes helper app (STRIPES-558)
 
 ## [1.3.0](https://github.com/folio-org/ui-inventory/tree/v1.3.0) (2018-09-27)
 [Full Changelog](https://github.com/folio-org/ui-inventory/compare/v1.2.1...v1.3.0)